Added wouldSend method for large message checking
The public wouldSend method returns the messages that would be sent if .say() or .notice() were called.
These both internally call _speak(), which splits longs lines and then sends. This could be refactored
so that the wouldSend method is used by the _speak method (after renaming). I.e. there could exist a
_splitMessages function that is used by both _speak and wouldSend to split messages.