parseMessage: allow . (dot), : (comma) and / (slash) in the nickname
Allow . (dot), : (colon) and / (slash) in the nickname for compatibility
with certain substandard IRC implementations (like the PSYC's built-in IRC
gate). However, only allow those characters when we are sure that we are not
taking a server hostname for a nick (otherwise any server hostname would get
misparsed as a nick without !user@host).