Make iconv dependency optional again.
Although iconv is listed in optionalDependencies, it's always required
in Client.prototype.convertEncoding function in lib/irc.js file,
even if neither "encoding" nor "encodingFallback" options are defined.

This error was introduced in 19daa5b2 when require('iconv') was
erroneously moved to the top of that function to avoid code duplication.

This change makes iconv module not required again when encoding and
encodingFallback options are not defined.

Signed-off-by: Oleg Girko <ol@infoserver.lv>